Issue #2416 Moved computation
